sync_and_async/pyproject.toml
[project]
name = "sync-and-async"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "env>=0.1.0",
    "openai-agents>=0.0.18",
    "python-dotenv>=1.1.0",
]
